# internal/api/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        example: descriptive error message
        type: string
    type: object
...
  api.syncStatusResponse:
    properties:
      conflict_count:
        example: 0
        type: integer
      conflict_note_ids:
        items:
          type: string
        type: array
      initial_sync_complete:
        example: "true"
        type: string
      last_push_at:
        example: "2025-01-15T10:25:00Z"
        type: string
      last_sync_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      queue_size:
        example: 3
        type: integer
    type: object
...
info:
  contact: {}
  description: API for reading and writing Bear notes via the hub.
  title: Bear Sync Hub API
  version: "1.0"
paths:
...
  /api/sync/status:
    get:
      description: Returns the current sync status including last sync times, queue
        size, and conflict info.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.syncStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sync status
      tags:
      - Sync
...
securityDefinitions:
  BearerAuth:
    description: 'Consumer Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"