# internal/api/sync_handler.go
package api

import (
	"context"
	"encoding/json"
	"io"
	"log/slog"
	"net/http"
	"os"
	"path/filepath"
	"time"

	"github.com/go-chi/chi/v5"

	"github.com/romancha/bear-sync/internal/models"
)

func (s *Server) syncPush(w http.ResponseWriter, r *http.Request) {
	var req models.SyncPushRequest
	if err := readJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid request body")
		return
	}

	// Collect attachment IDs to clean up before DB deletion removes the records.
	cleanupIDs := s.collectAttachmentCleanupIDs(r.Context(), req)

	if err := s.store.ProcessSyncPush(r.Context(), req); err != nil {
		writeInternalError(w, "failed to process sync push", err)
		return
	}

	// Clean up attachment files from disk after successful DB processing.
	s.cleanupAttachmentFiles(cleanupIDs)

	now := time.Now().UTC().Format(time.RFC3339)

	if err := s.store.SetSyncMeta(r.Context(), "last_push_at", now); err != nil {
		slog.Warn("failed to set last_push_at", "error", err)
	}

	// Store any metadata key-value pairs from the push request, skipping protected keys.
	for k, v := range req.Meta {
		if k == "last_push_at" {
			continue
		}
		if err := s.store.SetSyncMeta(r.Context(), k, v); err != nil {
			slog.Warn("failed to set sync meta", "key", k, "error", err)
		}
	}

	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

// collectAttachmentCleanupIDs gathers attachment IDs whose files should be removed from disk.
// This includes explicitly deleted attachments and those marked as permanently_deleted.
//
//nolint:gocritic // req is read-only, no need for pointer
func (s *Server) collectAttachmentCleanupIDs(ctx context.Context, req models.SyncPushRequest) []string {
	ids := make([]string, 0, len(req.DeletedAttachmentIDs))

	// Attachments in deleted_attachment_ids will be removed from DB — clean their files.
	for _, bearID := range req.DeletedAttachmentIDs {
		att, err := s.store.GetAttachmentByBearID(ctx, bearID)
		if err != nil || att == nil {
			continue
		}
		ids = append(ids, att.ID)
	}

	// Attachments pushed with permanently_deleted=1.
	for i := range req.Attachments {
		if req.Attachments[i].PermanentlyDeleted == 1 {
			if req.Attachments[i].BearID != nil && *req.Attachments[i].BearID != "" {
				att, err := s.store.GetAttachmentByBearID(ctx, *req.Attachments[i].BearID)
				if err != nil || att == nil {
					continue
				}
				ids = append(ids, att.ID)
			}
		}
	}

	return ids
}

// cleanupAttachmentFiles removes attachment directories from disk.
func (s *Server) cleanupAttachmentFiles(attachmentIDs []string) {
	for _, id := range attachmentIDs {
		dir := filepath.Join(s.attachmentsDir, id)
		if err := os.RemoveAll(dir); err != nil { //nolint:gosec // path from internal data
			slog.Warn("failed to remove attachment files", "attachment_id", id, "error", err)
		} else {
			slog.Debug("removed attachment files", "attachment_id", id)
		}
	}
}

// collectAddFileCleanupIDs gathers attachment IDs from add_file queue items
// that were successfully acked (now in "applied" status).
// Must be called AFTER AckQueueItems to avoid TOCTOU races.
func (s *Server) collectAddFileCleanupIDs(ctx context.Context, items []models.SyncAckItem) []string {
	ids := make([]string, 0, len(items))

	for i := range items {
		if items[i].Status != "applied" {
			continue
		}

		queueItem, err := s.store.GetQueueItem(ctx, items[i].QueueID)
		if err != nil || queueItem == nil || queueItem.Action != "add_file" {
			continue
		}

		// Only clean up if the ack was actually applied (status is now "applied").
		// If the ack was a no-op (item was reset to "pending" or already terminal),
		// the blob must remain for re-lease.
		if queueItem.Status != "applied" {
			continue
		}

		var payload struct {
			AttachmentID string `json:"attachment_id"`
		}
		if err := json.Unmarshal([]byte(queueItem.Payload), &payload); err != nil || payload.AttachmentID == "" {
			continue
		}

		ids = append(ids, payload.AttachmentID)
	}

	return ids
}

func (s *Server) syncQueue(w http.ResponseWriter, r *http.Request) {
	processingBy := r.URL.Query().Get("processing_by")
	if processingBy == "" {
		processingBy = "bridge"
	}

	items, err := s.store.LeaseQueueItems(r.Context(), processingBy, 5*time.Minute)
	if err != nil {
		writeInternalError(w, "failed to lease queue items", err)
		return
	}

	if items == nil {
		items = []models.WriteQueueItem{}
	}

	writeJSON(w, http.StatusOK, items)
}

func (s *Server) syncAck(w http.ResponseWriter, r *http.Request) {
	var req models.SyncAckRequest
	if err := readJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid request body")
		return
	}

	if err := s.store.AckQueueItems(r.Context(), req.Items); err != nil {
		writeInternalError(w, "failed to ack queue items", err)
		return
	}

	// Collect cleanup IDs after ack to avoid TOCTOU race: only clean up blobs
	// for items that were actually transitioned to "applied" by the ack above.
	// Use a detached context so cleanup isn't skipped if the client disconnects
	// after the ack commits but before we finish reading queue items.
	cleanupIDs := s.collectAddFileCleanupIDs(context.Background(), req.Items)
	s.cleanupAttachmentFiles(cleanupIDs)

	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

func (s *Server) syncUploadAttachment(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	attachment, err := s.store.GetAttachment(r.Context(), id)
	if err != nil {
		writeInternalError(w, "failed to get attachment", err)
		return
	}

	// Fallback: the bridge may send bear_id instead of hub ID (mapper generates ephemeral IDs).
	if attachment == nil {
		attachment, err = s.store.GetAttachmentByBearID(r.Context(), id)
		if err != nil {
			writeInternalError(w, "failed to get attachment", err)
			return
		}
	}

	if attachment == nil {
		writeError(w, http.StatusNotFound, "attachment not found")
		return
	}

	dir := filepath.Join(s.attachmentsDir, attachment.ID)

	if err := os.MkdirAll(dir, 0o750); err != nil { //nolint:gosec // path from DB, not user input
		writeInternalError(w, "failed to create attachment directory", err)
		return
	}

	filename := filepath.Base(attachment.Filename)
	if filename == "" || filename == "." || filename == "/" {
		filename = "file"
	}

	filePath := filepath.Join(dir, filename)

	f, err := os.Create(filePath) //nolint:gosec // path is constructed from internal data, not user input
	if err != nil {
		writeInternalError(w, "failed to create file", err)
		return
	}

	if _, err := io.Copy(f, r.Body); err != nil {
		f.Close()           //nolint:errcheck,gosec // closing before cleanup
		os.Remove(filePath) //nolint:errcheck,gosec // clean up partial file
		writeInternalError(w, "failed to write file", err)
		return
	}

	if err := f.Close(); err != nil {
		os.Remove(filePath) //nolint:errcheck,gosec // clean up on flush failure
		writeInternalError(w, "failed to finalize file", err)
		return
	}

	attachment.FilePath = filePath
	if updateErr := s.store.UpdateAttachment(r.Context(), attachment); updateErr != nil {
		os.Remove(filePath) //nolint:errcheck,gosec // best-effort cleanup on DB failure
		writeInternalError(w, "failed to update attachment path", updateErr)
		return
	}

	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

func (s *Server) syncDownloadAttachment(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	attachment, err := s.store.GetAttachment(r.Context(), id)
	if err != nil {
		writeInternalError(w, "failed to get attachment", err)
		return
	}

	// If found in DB, serve via the normal path.
	if attachment != nil {
		s.serveAttachmentFile(w, r, attachment)
		return
	}

	// Fall back to serving consumer-uploaded files (addFile writes to disk without a DB record).
	s.serveConsumerUploadedFile(w, r, id)
}

type syncStatusResponse struct {
	LastSyncAt          string   `json:"last_sync_at" example:"2025-01-15T10:30:00Z"`
	LastPushAt          string   `json:"last_push_at" example:"2025-01-15T10:25:00Z"`
	QueueSize           int      `json:"queue_size" example:"3"`
	InitialSyncComplete string   `json:"initial_sync_complete" example:"true"`
	ConflictCount       int      `json:"conflict_count" example:"0"`
	ConflictNoteIDs     []string `json:"conflict_note_ids,omitempty"`
}

// syncStatus godoc
// @Summary Get sync status
// @Description Returns the current sync status including last sync times, queue size, and conflict info.
// @Tags Sync
// @Produce json
// @Success 200 {object} syncStatusResponse
// @Failure 401 {object} ErrorResponse
// @Failure 403 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Security BearerAuth
// @Router /api/sync/status [get]
func (s *Server) syncStatus(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var resp syncStatusResponse

	resp.LastSyncAt, _ = s.store.GetSyncMeta(ctx, "last_sync_at")
	resp.LastPushAt, _ = s.store.GetSyncMeta(ctx, "last_push_at")
	resp.InitialSyncComplete, _ = s.store.GetSyncMeta(ctx, "initial_sync_complete")

	count, err := s.store.PendingQueueCount(ctx)
	if err == nil {
		resp.QueueSize = count
	}

	conflictCount, err := s.store.CountConflicts(ctx)
	if err == nil {
		resp.ConflictCount = conflictCount
	}

	if conflictCount > 0 {
		conflictIDs, err := s.store.ListConflictNoteIDs(ctx)
		if err == nil {
			resp.ConflictNoteIDs = conflictIDs
		}
	}

	writeJSON(w, http.StatusOK, resp)
}